\name{foil_colors_}
\alias{foil_colors_}
\title{Foil Colors}
\description{
Getter for colors from list of hex codes.
}
\usage{
foil_colors_(x)
}
\arguments{
  \item{x}{Name of color to get hex code for from list of hex codes}
}
\author{Tin Skoric}
\examples{
## get ultramarine hex code
foil_colors_("ultramarine")
