\name{foil_palette_}
\alias{foil_palette_}
\title{Foil Palette}
\description{
Getter for preset palettes containing colors from \code{foil_colors_}.
}
\usage{
foil_palette(x)
}
\arguments{
  \item{x}{Name of palette to get from list of palettes}
}
\author{Tin Skoric}
\examples{
## get flower palette
foil_colors_("flower")
