\name{theme_foil}
\alias{theme_foil}
\title{Foil Theme}
\description{
A theme with a dark background and white gridlines. The title is a play on words
of "tin foil," please laugh.
}
\usage{
theme_foil(base_size = 12, base_family = "")
}
\arguments{
  \item{base_size}{base font size}
  \item{base_family}{base font family}
}
\author{Tin Skoric}
\examples{
## add theme to plot
+ theme_foil()
